:recycle: add module paths
This may be too many layers of modules but we have to do something
because it wasn't building correctly the way we had it. We could drop
the outer-most modules and have them in one namespace.

According to the docs, the module layout should match the directory
layout because some generators use the directory structure. This commit
has the package name match the directory structure.