:bug: fix layout
I'm not sure why the `deprecated` directory was re-arranged the way it
was. This repo should simply be a directory tree of `.proto` files.
Nothing in it should be specific to a particular target because that
defeats the purpose of sharing the files.

I'll leave the `pom.xml` mess alone, as long as it stays in the
`deprecated` directory.