:recycle: Adjust how to support post trigger
This reorganization is a bit simpler than the previous and still allows
post triggers. Rather than having an arm event indicating the arming
goes at the end of collection, we just use the stop event. But that
means we can't use the `count` field, so that field was moved to the
`SampleEvent` message as a separate field. If 0, it means "infinite"
samples, otherwise it indicates an upper limit of how many points will
be sent.

Any acquisition service should treat a `StopEvent` that isn't `never`
and a `SampleEvent` with a `count` not zero as a post-trigger request;
that is, it keeps a circular buffer of the most recent `count` readings
and sends them when the `StopEvent` occurs.